requirements.txt
# Core dependencies
edge-tts==6.1.12
aiofiles==24.1.0
Pillow==10.4.0
requests==2.32.3
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
ffmpeg-python==0.2.0
python-dotenv==1.0.1
aiohttp==3.10.10
jinja2==3.1.4
schedule==1.2.2
pydantic==2.9.2
sqlite-utils==3.37
httpx==0.27.2
